Add a service implementing a stream interface over the MQTT protocol
The service uses a MQTT protocol dependency to subscribe to a topic
and forward all received MQTT traffic to a advertised stream interface.
All stream interface output will be in turn published to a selected
MQTT topic.

This allows us to easily implement simple services such as a remote
CLI console over the MQTT protocol.