feat: Reset before jumping to the app
Do not beat the wind here. Do not try to reset the MCU state before
jumping to the application firmware. It is **hard** to do properly.
Copy the relevant information to a struct residing in the .noinit
section to preserve data between boots and jump to the application
before the MCU state is severely changed.