fix: Increase XIP ELF header size to 0x200
On STM32 devices, the vector table must be aligned to the power
of 2 bigger than its size. Not doing this properly results in
bogus ISR handlers being run. Most of the newer STM32 devices
contain more handlers than 64 (bigger than 0x100 bytes). Increase
the header size to 0x200 to shift the vector table to be aligned
at the 0x200 boundary.